This repository consists VerilogHDL implementation of Advanced Encryption Standard (AES) with an edition to include dynamic s box
AES algorithm has been stated as secure against any attack but
increasing fast computing is making hackers to develop the cracks for
AES as well. Therefore to further increase the security of AES, i tried
to replace Standard static and fixed Substitution Box with a dynamic S
Box. Dynamicity is brought with the help of Input key. Static S box is
altered using the input key and the new generated s box is used for
encryption. Reverse steps goes for Decryption.

Presently, working on to test this design against Side Channel attacks
and would publish the results here.